pyproject.toml
[tool.poetry]
name = "titan"
version = "0.5.0"
readme = "README.md"
description = "TITAN platform API services"
homepage = "https://github.com/benhid/titan-platform/"
repository = "https://github.com/benhid/titan-platform/"
authors = ["Antonio Benítez Hidalgo <antoniobenitezhid@gmail.com>"]
classifiers=[
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python",
    "Typing :: Typed",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
]
include = [
    "LICENSE"
]

[tool.poetry.scripts]
titan = "titan.__main__:cli"

[tool.poetry.dependencies]
python = "^3.7"
pydantic = {extras = ["dotenv"], version = "^1.5.1"}
fastapi = "^0.61.2"
uvicorn = "^0.11.5"
pyjwt = "^1.7.1"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
python-multipart = "^0.0.5"
strconv = "^0.4.2"
httpx = "^0.15.4"
aiofiles = "^0.5.0"
aiocache = "^0.11.1"
motor = "^2.3.0"
pystardog = "^0.9.7"
rdflib = "^5.0.0"

[tool.poetry.dev-dependencies]
black = "^20.8.b1"
mypy = "^0.780"
isort = "^5.5.2"
pytest = "^6.1.0"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | _build
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
